Change type Product.price from Integer to Float

+++ app/models.py
@@ -25,7 +25,7 @@ class Product(Base):
 
     slug = Column(String, unique=True, index=True)
     description = Column(String)
-    price = Column(Integer)
+    price = Column(Float)
     image_url = Column(String)
     stock = Column(Integer)
     rating = Column(Float)
